Create the car entity class
We add the following dependencies to the pom.xml file and create the
following package that will name domain under the root package
com.packt.cardatabase. Next, we create our entity class, the name of the
entity class is Car.

We open the Car class and annotated with the @Entity. The Entity
annotation is imported from the javax.persistence package.

We add some fields to our class. The entity fields are mapped to
database table columns. The entity class must also contain a unique ID
that is used as primary key in the database.

The primary key is defined by using the @Id annotation. The
@GeneratedValue annotation defines that the ID is automatically
generated by the database. The AUTO type means that the JPA provider
select the best strategy for aparticular database. We can also create a
composite primary key by annotating multiple attributes with the @Id
annotation.

The database columns are named according to class field naming by
default. Use other naming convention, we can use the @Column annotation.

The table called car must be created in the database when we run the
application. To ensure that, we will add one new property to the
application.properties file.

spring.jpa.show-sql=true

To enable the H2 console, we add the following lines to the
application.properties file. The first setting enables the H2 console
and the second setting the endpoint that we can use to access the
console.

spring.h2.console.enabled=true
spring.h2.console.path=/h2-console

To access the H2 console we must to navinage to
localhost:8080/h2-console on the browser. We use jdbc:h2:mem:testdb as
the JDBC URL and leave the password field empty in the login window.

The Spring Boot Data JPA provides a CrudRepository interface for CRUD
operations.

We create our repository in the domain package. The interface is called
CarRepository. This interface extends the Spring Boot JPA CrudRepository
interface. The <Car, Long> type arguments define that this is the
repository for the Car entity class and the type of the ID field is
Long.

Let's use the Spring Boot CommandLineRunner interface. The
CaommandLineRunner interface allows us to execute additional code before
the application has fully started. We add demo data to our database.
CommandLineRunner is located inside the main class.

To inject our car repository into the main class to be able to save new
car objects to the database. We use @Autowired annotation to enable
dependency injection. The dependency injection allow us to pass
dependencies into an object. After we have injected the repository
class, we can use the CRUD methods it provides.

Finnally, we define our own queries in the Spring Data repositories. The
query must start with a prefix. After the prefix, you define the entity
class fields that are used in the query.