Create a one-to-many relationship
The one-to-many relationship can be added by using the @ManyToOne and
@OneToMany annotations. In the car entity class, which contains a
foreign key, you will define the relationship with the @ManyToOne
annotation. You should also add the getter and setter for the owner
field. It is recommended that you use FetchType.LAZY for all
associations. For the toMany relationShips, that is the default value,
but for the toOne relationships, you should define it. FetchType defines
the strategy for fetching data from the database. The value can be
either EAGER or LAZY. In our case, the lazy strategy means that when the
owner is fetched from the database, all the cars will be fetched
immediately with the owner.

In the owner entity site, the relationship is defined with the
@OneToMany annotation. The type of the field is List<Car>, because the
owner may have multiple cars. You can now add the getter and setter.

The @OneToMany annotation has two attribute that we are using. The
cascade attribute defines how cascading affects the entities. The ALL
attribute setting means that, if the owner is deleted, the cars linked
to that owner are deleted as well. The mappedBy="owner" attribute
setting tells us that the Car class has the owner field, which is the
fereign key for this relationship.

Now, we can add some owners to the database with CommandLineRunner.
Let's also modify the Car entity classes constructor and add an owner
object.

We first create two owner objects and save these to the database. In
order to save the owners, we also have to inject to OwnerRepository into
the main class. Then, we will connect the owners to the cars by using
the Car constructor.