Create a many-to-many relationships
An owner can have multiple cars and a car can have multiple owners, you
should use the @ManyToMany annotation. In a many-to-many relationship,
it is recommended what you use Set instead of List with hibernate.

When we run the application, there will be a new join table that is
created between the car and owner tables. The join table is defined by
using the @JoinTable annotation. With this annotation, we can set the
name of the join table and join columns.