Setting up PostgreSQL database
Now, will switch our database from H2 to PostgreSQL. The database tables
are still created automatically by JPA. But, before we run our
application, we have to create a database for it.

- We create the database via psql.

$ create database cardb;

We add a PostgreSQL dependency to the pom.xml file and remove the H2
dependency injection that we don't need anymore.

In the application.properties file, you will now define the database
connection. First, you will define the database's url, username,
password, and database driver class. The spring.jpa.generate-ddl setting
defines whether JPA should initialize the database(true/false).
The spring.jpa.hibernate.ddl-auto setting defines the behavior of the
database initialization. The possible values are none, validate, update,
create, and create-drop. create-drop means that the database is created
when an application starts and it is dropped when the application is
stopped. create-drop is also the default value if you don't define any.
The create value only creates the database when the application is
started. The update value creates the database and updates the schema if
it is changed.