Create a RESTful web service
Create a new packe under the com.packt.cardatabase package called web.

Create a new class under the com.packt.cardatabase.web called
CarController.

We open the CarController class and add the @RestController annotation
before the class definition. The @RestController annotation identifies
that this class will be the controller for the RESTful web service.

We add a new method inside our controller class. The method is annotated
with the @RequestMapping annotation, which defines the endpoint that the
method is mapped to. When a user navigates to the /cars	endpoint, the
getCars() method is executed.

The getCars() method returns all the car objects, which are then
marshaled to JSON objects by the Jackson library.

NOTE
By default, @RequestMapping handles all the HTTP method (GET, PUT, POST,
and more) request. You can define which method is accepted using the
following @RequestMapping("/cars", method=GET) parameter. Now, this
method handles only GET requests from the /cars endpoint.

To be able to return cars from the database, we have to inject our
CardRepository into the controller. Then, we can use the findAll()
method that the repository provides to fetch all cars.

Now, we are ready to run our application and navigate to
localhost:8080/cars. We can see that there is something wrong, and the
application seems to be in an infinite loop. This happens on account of
our one-to-many relationships between tha car and owner tables.
The car is serialized, and it contains an owner who is then serialized,
and that, in turn, contains cars that are the serialized and so on. To
avoid this, we have to add the @JsonIgnore annotation to the cars field
in the Owner class.