Create a new class called SecurityConfig
The @Configuration and @EnableWebSecurity annotations switch off the
default web security configuration, and we can define our own
configuration in this class. Inside the configure(HttpSecurity http)
method, we can define which in our application are secure and which are
not. We actually don't need this method yet because we can use the
default settings where all the endpoints are secured.

We add an in-memory users to our application SecurityConfig class. Also
we create an in-memory user with the username user and the password
password.

The use of in-memory users is good in the development phase, but the
real application should save the users in the database.