Validate users from the database
We created a new class call User in the domain package. We annotated the
User class with the @Entity annotation. We added the id, username,
password, and role class fields. Finally, add the constructor, gettters,
and setters. Set us all the fields to be nullable, and specify that the
username must be unique by using the @Column annotation.

We created a new class called UserRepository in the domain package. Set
us a query method, findByUsername. Next, we created a class that
implements the UserDetailsService interface that's provided by Spring
Security. Spring Security uses this for user authentication and
authorization. We created a new package in the root package called
service.

We created a new class called UserDetailServiceImpl in the service
package. We have to inject the UserRepository class into the
UserDetailServiceImpl class because that is needed to fetch the user
from the database when Spring Security handles authentication. The
loadByUsername method returns the UserDetails object, which is required
for authentication.

In our security configuration class, we have to stipulate that Spring
Security should use users from database instead of in-memory users.
Delete the userDetailsService() method from the SecurityConfig class to
disable in-memory users. Add a new configureGlobal method to enable
users from the database. We should never save the password as plain text
to the database. Therefore, we will define a password hashing
algorithm in the configureGlobal method. In this example, we are using
the BCrypt algorithm. This can be easily implemented with the Spring
Security BCryptPasswordEncoder class.

Finally, we can save a couple of test users to the database in our
CommandLineRunner. Open the CardatabaseApplication.java file and add the
following code at the beginning of the class to inject UserRepository
into the main class.