Create a new class called AuthenticationService
It's created the class in the service package. We define a few constants
EXPIRATIONTIME defines the expiration time of the token in milliseconds,
while SIGNINGKEY is an algorithm-specific signing key that's used to
digitally sign the JWT. Your should use a base64-encoded string to do
this. PREFIX defines the prefix of the token, and the Bearer schema is
typically used. The addToken method creates the token and adds it to
request's Authorization header. The signing key is encoded using the
SHA-512 algorithm. The method also adds Access-Control-Expose-Headers to
the header with the Authorization value. This is needed because we are
unable to access the Authorization header through a JavaScript frontend
by default. The getAuthontication method gets the token from the
response Authorization header using the parser() method provided by the
jjwt library.