Create a new class called LoginFilter
It's created the LoginFilter class in the root package that handles
POST request to the /login endpont. The LoginFilter class extends Spring
Security's AbstractAuthenticationProcessingFilter interface, which
requires that you set the authenticationManager property. Authentication
is performed by the attemptAuthentication method. If authentication is
successfu, the successfulAuthentication method is executed. This method
will then call the addToken method in our service class, and the token
will be added to the Authorization header.