Add a CORS filter
We added a CORS(Cross-Origin Resources Sharing) filter in our security
configuration class. This is needed for the frontend, which is sending
requests from the other origin. The CORS filter intercepts requests, and
if these are identified as cross-origin, it adds proper headers to the
request. For that, we will use Spring Security's CorsConfigurationSource
interface. In this example, we will allow all HTTP methods and
headers(using "*"). You can define the list of permissile origins,
methods, and headers here if you require a more finely graded
definition.