Use H2 database for running test
By default, Spring Boot uses an in-memory database for testing. We are
now using MariDB, but H2 can also be used for testing if we add the
following dependency to the pom.xml file. The scope defines that the H2
database will only be used for running tests; otherwise, the application
will use the MariaDB database.