Create a new class called CarRpositoryTest
It's created the class in the root test package. Instead of the
@SpringBootTest annotation, @DataJpaTest can be used if the test focuses
only on JPA components. When using this annotation, the H2 database,
Hibernate, and Spring Data are configured automatically for testing. SQL
logging is also turned on. The tests are transactional by default and
roll back at the end of the test case. TestEntityManager is used to
handle the persist entities and is designed to be used in testing.