Test the RESTful web service JWT authentication
To test the controllers or any endpoint that is exposed, we can use a
MockMvc object. By using the MockMvc object, the server is not started,
but the tests are performed in the layer where Spring handles HTTP
requests, and therefore it mocks the real situation. MockMvc provides
the perform method to send these requests. To test authentication, we
have to add credentials to request body. We will perform two requests
the first has the correct credentials, and we check that the status is
OK. The second request contains incorrect credentials and we check that
we get a 4XX HTTP error.